Ran 2to3 on ironpython tests.

+++ Tests/test_bigint.py
@@ -56,15 +56,15 @@ def test_misc():
 
 def test_hex_conversions():
     # Test hex conversions. CPython 2.5 uses capital L, lowercase letters a...f)
-    s = hex(27L)  # 0x1b
+    s = hex(27)  # 0x1b
     Assert(s == "0x1bL", "27L: Expect lowercase digits. Received: %s." % (s));
     
-    s = hex(-27L)
+    s = hex(-27)
     Assert(s == "-0x1bL", "-27L: Expect lowercase digits. Received: %s." % (s));
 
 def test_negative_misc():
     AssertError(ValueError, #"invalid literal for long() with base 10: ''",
-                lambda: long(''))
+                lambda: int(''))
 
 #--MAIN------------------------------------------------------------------------
 run_test(__name__)
